+++ HISTORY.rst
@@ -6,6 +6,12 @@ History
 
 GitHub Log
 ----------
+* Oct 8, 2020
+    - (by: sonofeft)
+        - version 1.1.19, added frozen flags to many methods
+        - removed depricated useFastLookup logic
+        - added tests for new frozen flags
+    
 * July 30, 2020
     - (by: sonofeft)
         - version 1.1.18, trying some python 3.8 DLL import fixes